Auto-authentication flow will break the redirect loop after the first try
There are cases when authentication worked on SSO side, but still fails when user is redirected back:
* 3rd party cookies are blocked
* User exists on SSO provider side, but not in Tiger

We don't want the script to redirect user to SSO and back endlessly. Yet, we do want auto-auth to run again after page reload.

JIRA: RAIL-4608
