+++ libs/sdk-ui-web-components/src/index.ts
@@ -11,9 +11,9 @@
  */
 
 import { CustomElementContext, getContext, setContext } from "./context";
-import { parseUrl } from "./parseUrl";
 import { Insight } from "./visualizations/Insight";
 import { Dashboard } from "./visualizations/Dashboard";
+import initializeAutoAuth from "./autoAuth";
 
 // Include styles async to use native link injection from MiniCssExtractPlugin
 import("./visualizations/components.css").catch((error) => {
@@ -21,74 +21,7 @@ import("./visualizations/components.css").catch((error) => {
     console.error("Failed to load analytics styles", error);
 });
 
-const authMethodsMap = {
-    sso: async function configureTigerSso(hostname: string) {
-        const {
-            default: tigerBackend,
-            ContextDeferredAuthProvider,
-            redirectToTigerAuthentication,
-        } = await import("./tigerBackend");
-
-        return tigerBackend({ hostname }).withAuthentication(
-            new ContextDeferredAuthProvider(redirectToTigerAuthentication),
-        );
-    },
-    bearSso: async function configureBearSso(hostname: string) {
-        const {
-            default: bearBackend,
-            ContextDeferredAuthProvider,
-            redirectToBearSsoAuthentication,
-        } = await import("./bearBackend");
-
-        return bearBackend({ hostname }).withAuthentication(
-            new ContextDeferredAuthProvider(redirectToBearSsoAuthentication),
-        );
-    },
-    bear: async function configureBearAuth(hostname: string) {
-        const {
-            default: bearBackend,
-            ContextDeferredAuthProvider,
-            redirectToBearAuthentication,
-        } = await import("./bearBackend");
-
-        return bearBackend({ hostname }).withAuthentication(
-            new ContextDeferredAuthProvider(redirectToBearAuthentication),
-        );
-    },
-};
-
-const timeout = (ms: number) => new Promise((res) => setTimeout(res, ms));
-
-const initializeAutoAuth = async () => {
-    // Detect current host based on the script's URL
-    const parsedUrl = parseUrl(import.meta.url);
-
-    const { hostname, authType, workspaceId } = parsedUrl;
-
-    const getBackend = authMethodsMap[authType];
-
-    // No auto-auth. Give a developer a margin of 5sec to define programmatic flow
-    if (!getBackend || !hostname) {
-        Promise.race([timeout(5000), getContext()]).then((res) => {
-            if (!res) {
-                // Resolved without a value === timeout
-                // eslint-disable-next-line no-console
-                console.warn(
-                    `Automatic authentication with ${
-                        hostname ?? "analytics server"
-                    } is not enabled. You can enable it by adding "?auth=sso" to your script URL or provide authentication context programmatically with "setContext" method.`,
-                );
-            }
-        });
-        return;
-    }
-
-    const backend = await getBackend(hostname);
-
-    setContext({ backend, workspaceId });
-};
-
-initializeAutoAuth().catch((error) => {
+initializeAutoAuth(import.meta.url).catch((error) => {
     // eslint-disable-next-line no-console
     console.error("Failed to configure automatic authentication flow", error);
 });
